requirements.txt
asgiref==3.4.1
Django==3.2.6
image==1.5.33
nibabel==3.2.1
numpy==1.21.2
packaging==21.0
Pillow==8.3.1
progress==1.6
pyparsing==2.4.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
